feat: update_relationship バリデーション・共有エピソード・Dashboard 改善 (v0.2.7)
## ego-mcp

### バグ修正
- update_relationship にフィールド名バリデーションとエイリアス解決を追加
  - _UPDATABLE_FIELDS ホワイトリストによる不正フィールド検出
  - trust → trust_level 等のエイリアス自動解決
  - inputSchema に enum 制約を追加
- dominant_tone 更新が emotional_baseline を通じて永続化されるように修正
- consider_them で baseline_tone を表示（observed_tone と分離）

### 新機能
- remember に shared_with / related_memories パラメータを追加
- shared_with 指定時に EpisodeStore.create() + RelationshipStore.add_shared_episode() を自動実行
- scaffold に shared_with のヒントを追加

### リファクタリング
- server.py を 11 モジュールに分割（public API 変更なし）
- memory.py を 4 モジュールに分割（public API 変更なし）

## Dashboard

### バグ修正
- Emotional state が latest_emotion フィールドで最新感情を正しく表示
- Activity feed が HTTP ポーリングフォールバック時もログを表示
- WebSocket 接続状態インジケータを追加

### 変更
- SessionSummary から latest latency パネルを削除（2 カラム化）

## ドキュメント
- CHANGELOG をルートに統合し ego-mcp/CHANGELOG.md を削除
