+++ dashboard/frontend/src/components/now/now-tab.tsx
@@ -3,22 +3,22 @@ import { AnomalyAlerts } from '@/components/now/anomaly-alerts'
 import { DesireRadar } from '@/components/now/desire-radar'
 import { EmotionStatus } from '@/components/now/emotion-status'
 import { SessionSummary } from '@/components/now/session-summary'
-import type { LogLine } from '@/hooks/use-dashboard-socket'
-import type { CurrentResponse } from '@/types'
+import type { CurrentResponse, LogLine } from '@/types'
 
 type NowTabProps = {
   current: CurrentResponse | null
   logLines: LogLine[]
+  connected: boolean
 }
 
-export const NowTab = ({ current, logLines }: NowTabProps) => (
+export const NowTab = ({ current, logLines, connected }: NowTabProps) => (
   <div className="space-y-4">
-    <SessionSummary current={current} />
+    <SessionSummary />
     <EmotionStatus current={current} />
     <div className="grid gap-4 lg:grid-cols-2">
       <DesireRadar current={current} />
       <div className="space-y-4">
-        <ActivityFeed logLines={logLines} />
+        <ActivityFeed logLines={logLines} connected={connected} />
         <AnomalyAlerts />
       </div>
     </div>
