+++ dashboard/frontend/src/hooks/use-dashboard-socket.ts
@@ -1,16 +1,7 @@
 import { useCallback, useEffect, useRef, useState } from 'react'
 
-import { fetchCurrent } from '@/api'
-import type { CurrentResponse } from '@/types'
-
-export type LogLine = {
-  ts: string
-  tool_name?: string
-  ok: boolean
-  level?: string
-  logger?: string
-  message?: string
-}
+import { fetchCurrent, fetchLogs } from '@/api'
+import type { CurrentResponse, LogLine } from '@/types'
 
 type DashboardSocketState = {
   current: CurrentResponse | null
@@ -25,6 +16,22 @@ const WS_URL =
 const RECONNECT_INTERVAL = 3_000
 const POLL_INTERVAL = 2_000
 const MAX_LOG_LINES = 50
+
+const lineKey = (line: LogLine) =>
+  `${line.ts}|${line.logger ?? ''}|${line.level ?? ''}|${line.message ?? ''}|${line.tool_name ?? ''}`
+
+const deduplicateAndMerge = (prev: LogLine[], incoming: LogLine[]) => {
+  if (incoming.length === 0) return prev
+  const next = [...prev]
+  const known = new Set(prev.map(lineKey))
+  for (const line of incoming) {
+    const key = lineKey(line)
+    if (known.has(key)) continue
+    known.add(key)
+    next.push(line)
+  }
+  return next.length > MAX_LOG_LINES ? next.slice(-MAX_LOG_LINES) : next
+}
 
 export const useDashboardSocket = (): DashboardSocketState => {
   const [current, setCurrent] = useState<CurrentResponse | null>(null)
@@ -42,6 +49,10 @@ export const useDashboardSocket = (): DashboardSocketState => {
       if (disposedRef.current) return
       const data = await fetchCurrent()
       if (!disposedRef.current) setCurrent(data)
+      const logs = await fetchLogs()
+      if (!disposedRef.current && logs.length > 0) {
+        setLogLines((prev) => deduplicateAndMerge(prev, logs))
+      }
     }
     void poll()
     pollTimerRef.current = setInterval(poll, POLL_INTERVAL)
@@ -80,52 +91,35 @@ export const useDashboardSocket = (): DashboardSocketState => {
         if (msg.type === 'current_snapshot' && msg.data) {
           setCurrent(msg.data as unknown as CurrentResponse)
         } else if (msg.type === 'log_line' && msg.data) {
-          setLogLines((prev) => {
-            const fields =
-              typeof msg.data?.fields === 'object' && msg.data.fields !== null
-                ? (msg.data.fields as Record<string, unknown>)
-                : {}
-            const level =
-              typeof msg.data?.level === 'string' ? msg.data.level : undefined
-            const message =
-              typeof msg.data?.message === 'string'
-                ? msg.data.message
+          const fields =
+            typeof msg.data?.fields === 'object' && msg.data.fields !== null
+              ? (msg.data.fields as Record<string, unknown>)
+              : {}
+          const level =
+            typeof msg.data?.level === 'string' ? msg.data.level : undefined
+          const message =
+            typeof msg.data?.message === 'string' ? msg.data.message : undefined
+          const toolName =
+            typeof msg.data?.tool_name === 'string'
+              ? msg.data.tool_name
+              : typeof fields.tool_name === 'string'
+                ? fields.tool_name
                 : undefined
-            const toolName =
-              typeof msg.data?.tool_name === 'string'
-                ? msg.data.tool_name
-                : typeof fields.tool_name === 'string'
-                  ? fields.tool_name
-                  : undefined
-            const line: LogLine = {
-              ts: (msg.data?.ts as string) ?? new Date().toISOString(),
-              tool_name: toolName,
-              ok:
-                typeof msg.data?.ok === 'boolean'
-                  ? msg.data.ok
-                  : !(level === 'ERROR' || message === 'Tool execution failed'),
-              level,
-              logger:
-                typeof msg.data?.logger === 'string'
-                  ? msg.data.logger
-                  : undefined,
-              message,
-            }
-            const last = prev[prev.length - 1]
-            if (
-              last &&
-              last.ts === line.ts &&
-              last.logger === line.logger &&
-              last.message === line.message &&
-              last.tool_name === line.tool_name
-            ) {
-              return prev
-            }
-            const next = [...prev, line]
-            return next.length > MAX_LOG_LINES
-              ? next.slice(-MAX_LOG_LINES)
-              : next
-          })
+          const line: LogLine = {
+            ts: (msg.data?.ts as string) ?? new Date().toISOString(),
+            tool_name: toolName,
+            ok:
+              typeof msg.data?.ok === 'boolean'
+                ? msg.data.ok
+                : !(level === 'ERROR' || message === 'Tool execution failed'),
+            level,
+            logger:
+              typeof msg.data?.logger === 'string'
+                ? msg.data.logger
+                : undefined,
+            message,
+          }
+          setLogLines((prev) => deduplicateAndMerge(prev, [line]))
         }
       }
