+++ dashboard/src/ego_dashboard/store.py
@@ -151,6 +151,7 @@ class TelemetryStore:
         if not self._events:
             return {
                 "latest": None,
+                "latest_emotion": None,
                 "tool_calls_per_min": 0,
                 "error_rate": 0.0,
                 "window_24h": {"tool_calls": 0, "error_rate": 0.0},
@@ -186,6 +187,17 @@ class TelemetryStore:
                 and emotion_source.emotion_intensity is not None
             ):
                 latest_payload["emotion_intensity"] = emotion_source.emotion_intensity
+        latest_emotion = None
+        if emotion_source is not None:
+            valence_raw = emotion_source.numeric_metrics.get("valence")
+            arousal_raw = emotion_source.numeric_metrics.get("arousal")
+            latest_emotion = {
+                "ts": emotion_source.ts.isoformat(),
+                "emotion_primary": emotion_source.emotion_primary,
+                "emotion_intensity": emotion_source.emotion_intensity,
+                "valence": (float(valence_raw) if isinstance(valence_raw, (int, float)) else None),
+                "arousal": (float(arousal_raw) if isinstance(arousal_raw, (int, float)) else None),
+            }
         log_window = [log for log in self._logs if log.ts >= window_start]
         log_window_24h = [log for log in self._logs if log.ts >= window_24h_start]
         invocation_count = sum(
@@ -216,6 +228,7 @@ class TelemetryStore:
         )
         return {
             "latest": latest_payload,
+            "latest_emotion": latest_emotion,
             "tool_calls_per_min": tool_calls_per_min,
             "error_rate": error_rate,
             "window_24h": {"tool_calls": tool_calls_24h, "error_rate": error_rate_24h},
