+++ ego-mcp/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "ego-mcp"
-version = "0.2.6"
+version = "0.2.7"
 description = "Cognitive scaffolding MCP server for LLM personality continuity"
 requires-python = ">=3.11"
 dependencies = [
